Fix bug in position(long) and tune performance tests

+++ src/main/java/net/openhft/chronicle/ExcerptAppender.java
@@ -20,7 +20,7 @@ package net.openhft.chronicle;
  * @author peter.lawrey
  */
 public interface ExcerptAppender extends Excerpt {
-    public void startExcerpt(int capacity);
+    public void startExcerpt(long capacity);
 
     public void roll();
